+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Support:
 * Support marshmallow-sqlalchemy>=0.29.0.
 * Test against Python 3.12.
 * Drop support for Python 3.7.
+* Add field ``Config`` for Flask configuration value.
 
 0.15.0 (2023-04-05)
 *******************